+++ inst/tinytest/ctrdata_isrctn.R
@@ -342,3 +342,18 @@ expect_message(
 expect_true(length(res) >= 5L)
 rm(res)
 
+
+#### ctrLoadQueryIntoDb documents ####
+
+tmpDir <- newTempDir()
+on.exit(unlink(tmpDir, recursive = TRUE), add = TRUE)
+
+expect_message(
+  ctrLoadQueryIntoDb(
+    queryterm = "https://www.isrctn.com/search?q=alzheimer",
+    con = dbc,
+    documents.path = tmpDir,
+    documents.regexp = ".*"
+  ),
+  "Newly saved [0-9]+ document[(]s[)] for [0-9]+ trial"
+)